+++ file_dir.h
@@ -4,7 +4,7 @@ typedef struct file_list
 {
 	char *filename;
 	struct file_list *next;
-        int number;
+        short number;
 } F_set;
 
 extern void get_file_path(const char *path, const char *file_name,  char *file_path);
